fix errors with new repo write traits

+++ mm_store/src/archive.rs
@@ -6,7 +6,7 @@ use crate::Checksum;
 
 #[derive(Debug, Serialize, Deserialize, Type)]
 pub struct ThinArchiveEntry {
-    offset: usize,
+    offset: u64,
     compressed_file: Checksum,
     uncompressed_file: Checksum,
     compression_method: CompressionMethod,
